backend/requirements.txt
fastapi==0.115.0
uvicorn==0.32.0
pydantic==2.10.0
pydantic-settings==2.6.1
pandas==2.2.3
numpy==2.1.0
requests==2.31.0
google-play-scraper==1.2.4
beautifulsoup4==4.12.2
jsonschema==4.20.0
python-multipart==0.0.6
scipy==1.14.1
scikit-learn==1.5.2
altair==5.1.2
plotly==5.18.0
jinja2==3.1.4
markupsafe==3.0.2
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-dotenv==1.0.1
reportlab==4.2.5
APScheduler==3.10.4
slowapi==0.1.9
duckduckgo-search==6.4.2
spacy==3.7.0
json-repair==0.30.0
gunicorn==21.2.0
bcrypt==4.0.1
celery==5.4.0
redis==5.1.1
psycopg2-binary==2.9.9
pgvector==0.3.4
